Copy events to main calling module.
This is a test, as it appears that events, although triggered are not decodable from the metadata of the calling module. This causes a display error in the UI.

What is being done here is that event declarations are copied over from their original module (duplicated) so that they can be decoded in the UI.

This may not work!