Reposition automatic generation of prefunding hash to open API
This change fixes a bug where the generation of the prefunding hash was independent from the calling code. This caused issues when used with traits.

The hash is now only generated when called directly from the public api, otherwise it uses an already known hash supplied from an external module.